acpi: expose data structurs and functions of BIOS linker loader
Expose some data structures and functions of BIOS linker loader
which will be used by later commits.
